+++ src/core/EditorShell.tsx
@@ -8,14 +8,22 @@ import { TablePlugin as LexicalTablePlugin } from "@lexical/react/LexicalTablePl
 import LexicalErrorBoundary from "@lexical/react/LexicalErrorBoundary";
 import SlashCommandPlugin from "../plugins/SlashCommandPlugin";
 import ToolbarPlugin from "../plugins/ToolbarPlugin";
-import ImageUploadPlugin from "../plugins/ImageUploadPlugin";
+import {ImageUploadPlugin} from "../plugins/ImageUploadPlugin";
 import LinkPlugin from "../plugins/LinkPlugin";
 import SectionEditorPlugin from "../plugins/SectionEditorPlugin";
 import EmbedPlugin from "../plugins/EmbedPlugin";
 import TablePlugin from "../plugins/TablePlugin";
 import "../styles/editor.css";
 
-export default function EditorShell({ onChange }: { onChange?: (state: any) => void }) {
+export function EditorShell({
+  onChange,
+  onImageAdded,
+  useBase64Url
+}: {
+  onChange?: (state: any) => void;
+  onImageAdded?: (file: File) => Promise<string>;
+  useBase64Url?: boolean;
+}) {
   return (
     <div className="cms-editor-shell">
       <ToolbarPlugin />
@@ -29,7 +37,7 @@ export default function EditorShell({ onChange }: { onChange?: (state: any) => v
       <LexicalLinkPlugin />
       <LexicalTablePlugin />
       <SlashCommandPlugin />
-      <ImageUploadPlugin />
+      <ImageUploadPlugin onImageAdded={onImageAdded} useBase64Url={useBase64Url} />
       <LinkPlugin />
       <SectionEditorPlugin />
       <EmbedPlugin />